+++ visionforge-core/src/commonMain/kotlin/hep/dataforge/vision/VisionGroup.kt
@@ -58,7 +58,7 @@ public operator fun VisionGroup.iterator(): Iterator<Vision> = children.values.i
 public val VisionGroup.isEmpty: Boolean get() = this.children.isEmpty()
 
 public interface VisionContainerBuilder<in V : Vision> {
-    public operator fun set(name: Name, child: V?)
+    public operator fun set(name: Name?, child: V?)
 }
 
 /**
@@ -79,6 +79,6 @@ public operator fun <V : Vision> VisionContainer<V>.get(str: String): V? = get(s
 public operator fun <V : Vision> VisionContainerBuilder<V>.set(token: NameToken, child: V?): Unit =
     set(token.asName(), child)
 
-public operator fun <V : Vision> VisionContainerBuilder<V>.set(key: String, child: V?): Unit = set(key.toName(), child)
+public operator fun <V : Vision> VisionContainerBuilder<V>.set(key: String?, child: V?): Unit = set(key?.toName(), child)
 
 public fun MutableVisionGroup.removeAll(): Unit = children.keys.map { it.asName() }.forEach { this[it] = null }