+++ visionforge-solid/src/commonMain/kotlin/hep/dataforge/vision/solid/SolidGroup.kt
@@ -83,10 +83,9 @@ public fun SolidGroup(block: SolidGroup.() -> Unit): SolidGroup {
 
 @VisionBuilder
 public fun VisionContainerBuilder<Vision>.group(
-    name: Name = Name.EMPTY,
+    name: Name? = null,
     action: SolidGroup.() -> Unit = {},
-): SolidGroup =
-    SolidGroup().apply(action).also { set(name, it) }
+): SolidGroup = SolidGroup().apply(action).also { set(name, it) }
 
 /**
  * Define a group with given [name], attach it to this parent and return it.